+++ src/ebus_sdk/__init__.py
@@ -43,7 +43,7 @@ from .property import (
 # MQTT client
 from ebus_mqtt_client import MqttClient
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 __all__ = [
     # Homie classes